Introduce abstract UuidId
The system will utilize uuids to identify its entities. To represent these uuids in
our code, we should use value objects.

To avoid repeating ourselves, we introduce a common layer that holds helpers used
throughout the code.

In this common layer we'll create an abstract uuid id class that contains the
necessary code to create, validate, and read the uuid id value object.