Add doctrine annotations to entities
While this violates the clean architecture principle it is a trade-off I'm willing to take.
Storing the database definition in config files in the infrastructure layer is a possibility but
would in practice be very cumbersome and prone to errors when updating the entities.

Because there is no standard way for Doctrine to hydrate into value objects the value objects are
deserialized to a string. The getters will serialize them back into value objects.