📝 Add initial migration
The first migration to create tables for both the article and source entity.

For now the migrations are stored in the root of the project. Depending on if
we want to persist data in a manner that does not share a kernel this will have
to change in the future.