Change purpose of create article command
In the current domain there is no need for a command that just creates
an article. All created articles will come from external sources. These
sources can update their articles whenever. Since we don't care about
keeping a log of these changes we will update the article if
the external feed pushes an update on one of their items.

Here we refactor the 'create article' command to update an article if
its url is already coupled to an article in our system.