Support updating an article with an ORM in mind
We made the mistake to create a new object when an article
should be updated. This results in the ORM tracker losing the link with
the row that is already stored in the database. Thus resulting in a
newly created entry in the database which fails of the unique
constraint on the id.

Here we refactor the handler to accommodate the ORM with keeping the
link to the existing object intact.