Introduce event bus
Here we introduce a new bus next to the existing command bus.
Technically the key difference between the event-
and commandbus is that the commandbus only has one handler
(one-to-one) and a event bus can have multiple (one-to-many).

The purpose of the event bus is fastly different then the
command bus. If a command is dispatched on the command bus we
expect a handler to do the action the command describes. With
the event bus we dispatch an event (mostly for telling the
system we did a thing) and we don't really care who does
something with it.