Refactor feed fetch flow
The current situation is not really scalable nor are the concerns
properly seperated. The CLI command defined in the infrastructure layer
fetched the feed, this should not be its responsibility. After fetching
the feed it calls the command x amount of times to create an article
based on the fetched feed. This is all done in a single lifecycle on a
single thread.

This refactor addresses both the scalability problem and the separating
of concerns. A new command is introduced with the sole responsibility of
fetching a single feed. For each item in the fetched feed this command
dispatches an event. A listener acts on these events and calls the
UpsertArticleCommand.

Now both the fetching of feeds and the upserting of articles are done
in smaller single idempotent commands. In the future these commands can
run asynchronously, by putting them in a queue, to allow horizontal
scaling.