🐛 Source should be unique for each feedProvider
The tagged locator uses the source as an index to retrieve the
corresponding feedprovider. If there is a source that points to multiple
services the locator fetches the first one it encounters. This results
in the second feed not being fetched.

AFAIK Symfony does not provide a locator that can fetch multiple
services by index.

There are multiple solutions to this problem:

- We could have the php.watch feed provider fetch both feeds.
- We could keep both the providers and give them a unique source.

In this commit we decided on the latter. We introduce two new sources,
php.watch-changes and php.watch-news.