👮 Prevent feed providers with duplicate source
The feed provider source should be unique for each class that extends
the FeedProvider interface.

In this commit we add a compiler pass who guards against violations of
this rule. The container won't compile if a developer picks a source
name that is already taken by another implementation of the
feed provider interface.