Move updating of article to entity itself
Currently, the command handler is in control of updating the state of
the article entity. While this works it results in a fragmented domain
layer. The domain layer itself can't control it's input, it has to rely
on the application layer to interpret the entity correctly.

In this commit we move the business logic associated with updating the
article to the entity itself.