Send events from the domain itself
Currently, the application layer is responsible for sending events that
are actually more related to the domain. For example, the `UpsertArticleHandler`
publishes an event when an article is created.

In this commit we move the publishing of domain events to the entity
itself. By adding a doctrine listener we make sure the events are only
emitted when the entity is committed to the database.