Use a generic FeedParser
Creating a new FeedProvider for each feed becomes a bit cumbersome.
The goal is to add a bunch of feeds, and it would be desired to do this
without having to do much work.

In this commit we introduce a generic rss parser, SimplePie. SimplePie
is able to parse any type of rss feed into a generic item. This reduces
the manual work we have to do for each provider.