Introduce feeditem normalizers
These normalizers can be used to straighten out the outputs from all the different
feed sources.

We are used to doing this in the feedprovider itself, but since we want to remove these
we need a general soltuion.

In this commit we introduce two normalizers, one that removes html tags and another one
that shortens the summary.