Add normalizers to feed-to-article chain
We have a couple of normalizers to 'clean' the feed before it is useful
in our domain. These are not applied in the new way of fetching feeds.

In this commit we add a command to the chain that for each feed item
applies the normalizers.

The chain now looks as follows:

Fetch feed -> normalize each feed item -> upsert article per feed item